Refactor gene data loading and Solr update scripts
- Updated `gene_name.py` to correct import order for `Name` model.
- Enhanced `main.py` in data loading to include detailed docstrings, error handling, and a database dump feature after processing.
- Improved `main.py` in data update to add comprehensive docstrings for clarity and maintainability.
- Modified `update.sh` to log data transfer operations and handle batch file renaming after successful loading.
- Updated `requirements.txt` to clarify extra requirements for data loading.

+++ bin/update.sh
@@ -1,5 +1,17 @@
+# Do DB data update
+# Check if batch.csv exists
+echo "Transfer data to DB" 2>&1 | tee output.log
+if [ -f /usr/src/app/input/batch.csv ]; then
+    python /usr/src/app/bin/data-load/main.py --file /usr/src/app/input/batch.csv || { echo "Data load failed" >&2; exit 1; }
+    TODAY=$(date +%Y-%m-%d)
+    echo "Data loaded successfully on $TODAY" 2>&1 | tee -a output.log
+    mv /usr/src/app/input/batch.csv /usr/src/app/input/batch_$TODAY.csv
+fi
+echo "Done" 2>&1 | tee output.log
+
+# Do Solr data update
 rm -f /usr/src/app/output/done.flag solr.json
 echo "Transfer data to solr" 2>&1 | tee output.log
-python ./data-update/main.py --clear 2>&1 | tee -a output.log
+python /usr/src/app/bin/data-update/main.py --clear 2>&1 | tee -a output.log
 echo "Transfer data to dump file" 2>&1 | tee -a output.log
-python ./data-update/main.py --dump 2>&1 | tee -a output.log
+python /usr/src/app/bin/data-update/main.py --dump 2>&1 | tee -a output.log
